# APICL.Client/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "APICL.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Audio/audios/{guid}/info": {
      "get": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AudioObjInfo"
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/Image/images/{guid}/image64/{format}": {
      "get": {
        "tags": [
          "Image"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "format",
            "required": true,
            "type": "string",
            "default": "bmp"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ImageData"
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/status": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OpenClServiceInfo"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/devices": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OpenClDeviceInfo"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/devices/{deviceId}/initialize": {
      "post": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 2
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/OpenClServiceInfo"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/dispose": {
      "delete": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OpenClServiceInfo"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/usage": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OpenClUsageInfo"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/memory": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OpenClMemoryInfo"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/kernels/{filter}": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "filter",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OpenClKernelInfo"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/executeMandelbrot/{kernel}/{version}/{width}/{height}/{zoom}/{x}/{y}/{coeff}/{r}/{g}/{b}/{copyGuid}/{allowTempSession}": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "kernel",
            "required": true,
            "type": "string",
            "default": "mandelbrotPrecise"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string",
            "default": "01"
          },
          {
            "in": "path",
            "name": "width",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 1920
          },
          {
            "in": "path",
            "name": "height",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 1080
          },
          {
            "in": "path",
            "name": "zoom",
            "required": true,
            "type": "number",
            "format": "double",
            "default": 1
          },
          {
            "in": "path",
            "name": "x",
            "required": true,
            "type": "number",
            "format": "double",
            "default": 0
          },
          {
            "in": "path",
            "name": "y",
            "required": true,
            "type": "number",
            "format": "double",
            "default": 0
          },
          {
            "in": "path",
            "name": "coeff",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 8
          },
          {
            "in": "path",
            "name": "r",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          {
            "in": "path",
            "name": "g",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          {
            "in": "path",
            "name": "b",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          {
            "in": "path",
            "name": "copyGuid",
            "required": true,
            "type": "boolean",
            "default": true
          },
          {
            "in": "path",
            "name": "allowTempSession",
            "required": true,
            "type": "boolean",
            "default": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ImageObjInfo"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/OpenCl/executeTimestretch/{guid}/{kernel}/{version}/{factor}/{chunkSize}/{overlap}/{copyGuid}/{allowTempSession}": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "kernel",
            "required": true,
            "type": "string",
            "default": "timestretch_double"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string",
            "default": "03"
          },
          {
            "in": "path",
            "name": "factor",
            "required": true,
            "type": "number",
            "format": "double",
            "default": 0.8
          },
          {
            "in": "path",
            "name": "chunkSize",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 16384
          },
          {
            "in": "path",
            "name": "overlap",
            "required": true,
            "type": "number",
            "format": "float",
            "default": 0.5
          },
          {
            "in": "path",
            "name": "copyGuid",
            "required": true,
            "type": "boolean",
            "default": true
          },
          {
            "in": "path",
            "name": "allowTempSession",
            "required": true,
            "type": "boolean",
            "default": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/AudioObjInfo"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "AudioObjInfo": {
      "type": "object",
      "properties": {
        "guid": {
          "format": "uuid",
          "type": "string"
        },
        "filepath": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "samplerate": {
          "format": "int32",
          "type": "integer"
        },
        "bitdepth": {
          "format": "int32",
          "type": "integer"
        },
        "channels": {
          "format": "int32",
          "type": "integer"
        },
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "memoryLocation": {
          "type": "string"
        },
        "chunkSize": {
          "format": "int32",
          "type": "integer"
        },
        "overlapSize": {
          "format": "int32",
          "type": "integer"
        },
        "form": {
          "type": "string"
        },
        "bpm": {
          "format": "float",
          "type": "number"
        },
        "playingString": {
          "type": "string"
        },
        "sizeString": {
          "type": "string"
        },
        "durationTimeSpan": {
          "format": "date-span",
          "type": "string",
          "readOnly": true
        },
        "durationString": {
          "type": "string"
        },
        "lastProcessingTime": {
          "type": "string"
        },
        "lastLoadingTime": {
          "type": "string"
        },
        "entry": {
          "type": "string"
        },
        "errorInfo": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ImageData": {
      "type": "object",
      "properties": {
        "guid": {
          "format": "uuid",
          "type": "string"
        },
        "base64": {
          "type": "string"
        },
        "width": {
          "format": "int32",
          "type": "integer"
        },
        "height": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "ImageObjInfo": {
      "type": "object",
      "properties": {
        "guid": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        },
        "height": {
          "format": "int32",
          "type": "integer"
        },
        "width": {
          "format": "int32",
          "type": "integer"
        },
        "bitdepth": {
          "format": "int32",
          "type": "integer"
        },
        "channels": {
          "format": "int32",
          "type": "integer"
        },
        "sizeInBytes": {
          "format": "int64",
          "type": "integer"
        },
        "pointer": {
          "format": "int64",
          "type": "integer"
        },
        "onHost": {
          "type": "boolean"
        },
        "entry": {
          "type": "string"
        },
        "lastProcessingTime": {
          "type": "string"
        },
        "lastLoadingTime": {
          "type": "string"
        },
        "errorInfo": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OpenClDeviceInfo": {
      "type": "object",
      "properties": {
        "deviceId": {
          "format": "int32",
          "type": "integer"
        },
        "deviceName": {
          "type": "string"
        },
        "deviceType": {
          "type": "string"
        },
        "platformName": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OpenClKernelInfo": {
      "type": "object",
      "properties": {
        "index": {
          "format": "int32",
          "type": "integer"
        },
        "filePath": {
          "type": "string"
        },
        "functionName": {
          "type": "string"
        },
        "argumentNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "argumentsCount": {
          "format": "int32",
          "type": "integer"
        },
        "argumentTypeNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "inputPointerTypeName": {
          "type": "string"
        },
        "outputPointerTypeName": {
          "type": "string"
        },
        "mediaType": {
          "type": "string"
        },
        "entry": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OpenClMemoryInfo": {
      "type": "object",
      "properties": {
        "pointers": {
          "type": "array",
          "items": {
            "format": "int64",
            "type": "integer"
          }
        },
        "lengths": {
          "type": "array",
          "items": {
            "format": "int64",
            "type": "integer"
          }
        },
        "indexPointer": {
          "format": "int64",
          "type": "integer"
        },
        "indexLength": {
          "format": "int64",
          "type": "integer"
        },
        "count": {
          "format": "int64",
          "type": "integer"
        },
        "totalLength": {
          "format": "int64",
          "type": "integer"
        },
        "dataTypeSize": {
          "format": "int32",
          "type": "integer"
        },
        "dataTypeName": {
          "type": "string"
        },
        "totalSizeBytes": {
          "format": "int64",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "OpenClServiceInfo": {
      "type": "object",
      "properties": {
        "deviceId": {
          "format": "int32",
          "type": "integer"
        },
        "deviceName": {
          "type": "string"
        },
        "platformName": {
          "type": "string"
        },
        "initialized": {
          "type": "boolean"
        },
        "status": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OpenClUsageInfo": {
      "type": "object",
      "properties": {
        "totalMemory": {
          "format": "int64",
          "type": "integer"
        },
        "usedMemory": {
          "format": "int64",
          "type": "integer"
        },
        "freeMemory": {
          "format": "int64",
          "type": "integer"
        },
        "usagePercentage": {
          "format": "float",
          "type": "number"
        },
        "sizeUnit": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    }
  }
}